refactor: ロジック部分の階層変更、アプデに伴う型修正(#10)
