+++ src/domains/getQuoteList/index.ts
@@ -1,0 +1,3 @@
+import getQuoteList from './getQuoteList';
+
+export default getQuoteList;
