+++ src/domains/getRandomQuote/index.ts
@@ -1,0 +1,3 @@
+import getRandomQuote from './getRandomQuote';
+
+export default getRandomQuote;
