+++ src/hooks/quote/useGetQuoteListInfiniteQuery/index.ts
@@ -1,0 +1,3 @@
+import useGetQuoteListInfiniteQuery from './useGetQuoteListInfiniteQuey';
+
+export default useGetQuoteListInfiniteQuery;
