+++ src/hooks/util/useIntersectionObserver/index.ts
@@ -1,0 +1,3 @@
+import useIntersectionObserver from './useIntersectionObserver';
+
+export default useIntersectionObserver;
